reset back to previous green 2b9dabdaa86bda59678063fd024af8f65d612dc7 with history.

+++ support/java/src/android/support/v8/renderscript/ScriptIntrinsic.java
@@ -25,7 +25,7 @@ package android.support.v8.renderscript;
  * Not intended for direct use.
  **/
 public abstract class ScriptIntrinsic extends Script {
-    ScriptIntrinsic(long id, RenderScript rs) {
+    ScriptIntrinsic(int id, RenderScript rs) {
         super(id, rs);
     }
 }
